pyproject.toml
[project]
name = "voice-notes-bot"
version = "0.1.0"
description = "Telegram bot for converting voice note files to voice messages"
authors = [
    {name = "Peter Minin", email = "peterminin@gmail.com"},
]
dependencies = [
    "python-telegram-bot~=21.3",
    "ffmpeg-python~=0.2.0",
]
requires-python = "~=3.11"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
run_bot = "python -m src.voice_notes_bot.bot ./local/secrets/bot.json ./local/config.json ./local/state/"
